fixing the input func

+++ src/ExcavationCracks.cpp
@@ -83,7 +83,7 @@ void stiffness_bb(float& aks_bb, float& akn_bb, float& phi_bb, float& coh_bb, fl
         float x2 = xc + ac * cosbeta;   //x2 = 0.106666669
         float y2 = yc + ac * sinbeta;   //y2 = 0.0533333346
         int  en;
-        float thr = 0.0003; //6e-4;
+        float thr =  0.0003; //6e-4;
         if (flagB == 0)
             en = numbe - 1;
         else 
@@ -97,8 +97,8 @@ void stiffness_bb(float& aks_bb, float& akn_bb, float& phi_bb, float& coh_bb, fl
                 if (be.kod != 5) continue;
             }
           
-            float dist1 = sqrtf(powf(x1 - be.xm, 2) + powf(y1 - be.ym, 2));
-            float dist2 = sqrtf(powf(x2 - be.xm, 2) + powf(y2 - be.ym, 2));
+            float dist1 = sqrt(powf(x1 - be.xm, 2) + powf(y1 - be.ym, 2));
+            float dist2 = sqrt(powf(x2 - be.xm, 2) + powf(y2 - be.ym, 2));
 
             
             if (min(dist1, dist2) - thr <= factors.tolerance * max(ac, be.a)  )
@@ -109,7 +109,7 @@ void stiffness_bb(float& aks_bb, float& akn_bb, float& phi_bb, float& coh_bb, fl
 
             if (symm.ksym == 1 || symm.ksym == 4) 
             {
-                dist1 = sqrtf(powf(x1 - (2.0 * symm.xsym - be.xm), 2) + 
+                dist1 = sqrt(powf(x1 - (2.0 * symm.xsym - be.xm), 2) + 
                     powf(y1 - be.ym, 2));
                 dist2 = sqrtf(powf(x2 - (2.0 * symm.xsym - be.xm), 2)+
                     powf(y2 - be.ym, 2));
@@ -123,8 +123,8 @@ void stiffness_bb(float& aks_bb, float& akn_bb, float& phi_bb, float& coh_bb, fl
 
             if (symm.ksym == 2 || symm.ksym == 4) 
             {
-                dist1 = sqrtf(powf(x1 - be.xm, 2) + powf(y1 - (2.0 * symm.ysym - be.ym), 2));
-                dist2 = sqrtf(pow(x2 - be.xm, 2) + powf(y2 - (2.0 * symm.ysym - be.ym), 2));
+                dist1 = sqrt(powf(x1 - be.xm, 2) + powf(y1 - (2.0 * symm.ysym - be.ym), 2));
+                dist2 = sqrt(pow(x2 - be.xm, 2) + powf(y2 - (2.0 * symm.ysym - be.ym), 2));
                 if (min(dist1, dist2) <= factors.tolerance * max(ac, be.a))
                 {
                     //legal = 0;
@@ -134,8 +134,8 @@ void stiffness_bb(float& aks_bb, float& akn_bb, float& phi_bb, float& coh_bb, fl
 
             if (symm.ksym == 3 || symm.ksym == 4) 
             {
-                dist1 = sqrtf(powf(x1 - (2.0 * symm.xsym - be.xm), 2) + powf(y1 - (2.0 * symm.ysym - be.ym), 2));
-                dist2 = sqrtf(powf(x2 - (2.0 * symm.xsym - be.xm), 2) + powf(y2 - (2.0 * symm.ysym - be.ym), 2));
+                dist1 = sqrt(powf(x1 - (2.0 * symm.xsym - be.xm), 2) + powf(y1 - (2.0 * symm.ysym - be.ym), 2));
+                dist2 = sqrt(powf(x2 - (2.0 * symm.xsym - be.xm), 2) + powf(y2 - (2.0 * symm.ysym - be.ym), 2));
                 if (min(dist1, dist2) <= factors.tolerance * max(ac, be.a) )
                 {
                     //legal = 0;
